upgrade version v1

+++ contracts/external/INarwhalswapRouter02.sol
@@ -2,45 +2,51 @@
 
 pragma solidity >=0.6.2;
 
-import './INarwhalswapRouter01.sol';
+import "./INarwhalswapRouter01.sol";
 
 interface INarwhalswapRouter02 is INarwhalswapRouter01 {
     function removeLiquidityBNBSupportingFeeOnTransferTokens(
         address token,
-        uint liquidity,
-        uint amountTokenMin,
-        uint amountBNBMin,
+        uint256 liquidity,
+        uint256 amountTokenMin,
+        uint256 amountBNBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountBNB);
+        uint256 deadline
+    ) external returns (uint256 amountBNB);
+
     function removeLiquidityBNBWithPermitSupportingFeeOnTransferTokens(
         address token,
-        uint liquidity,
-        uint amountTokenMin,
-        uint amountBNBMin,
+        uint256 liquidity,
+        uint256 amountTokenMin,
+        uint256 amountBNBMin,
         address to,
-        uint deadline,
-        bool approveMax, uint8 v, bytes32 r, bytes32 s
-    ) external returns (uint amountBNB);
+        uint256 deadline,
+        bool approveMax,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external returns (uint256 amountBNB);
 
     function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
+
     function swapExactBNBForTokensSupportingFeeOnTransferTokens(
-        uint amountOutMin,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external payable;
+
     function swapExactTokensForBNBSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
 }
